rename strategy->strategies
