+++ src/fsrs/index.ts
@@ -23,7 +23,7 @@ export { State, Rating } from './models'
 
 export * from './convert'
 
-export * from './strategy'
+export * from './strategies'
 export * from './abstract_scheduler'
 export * from './impl/basic_scheduler'
 export * from './impl/long_term_scheduler'
